backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flower==2.0.1
fonttools==4.50.0
greenlet==3.0.3
httplib2==0.22.0
humanize==4.9.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
